Added ProductCard Component, removed unused assets, moved Products.jsx to components
