+++ frontend/src/App.jsx
@@ -1,8 +1,5 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from './assets/vite.svg'
-import heroImg from './assets/hero.png'
-import Products from "./Products";
+import Products from "./components/Products";
 
 function App() {
   return (
